requirements.txt
seaborn
pandas
numpy
keras
matplotlib
hyperopt
gpustat
tensorflow-gpu==1.14